Added Dependency Injection

+++ WpfMvvmProject/WpfMvvmProject/Model/PlayerStats.cs
@@ -9,7 +9,7 @@ namespace WpfMvvmProject.Model;
 public class PlayerStats
 {
     public string? Name { get; set; }
-    public int PointsPerGameAvg { get; set; }
+    public double PointsPerGameAvg { get; set; }
 
     public int CareerPoints { get; set; }
